# internal/api/openapi.yaml
basePath: /api
definitions:
...
  api.appProxyEntry:
    properties:
      port:
        type: integer
      url:
        type: string
    type: object
...
  api.createProjectRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
  api.errorResponse:
    properties:
      error:
        type: string
    type: object
...
  api.projectResponse:
    properties:
      appCount:
        type: integer
      createdAt:
        type: string
      description:
        type: string
      name:
        type: string
      namespace:
        type: string
      phase:
        $ref: '#/definitions/v1alpha1.ProjectPhase'
    type: object
...
  v1alpha1.ProjectPhase:
    enum:
    - Pending
    - Ready
    - Terminating
    - Failed
    type: string
    x-enum-varnames:
    - ProjectPhasePending
    - ProjectPhaseReady
    - ProjectPhaseTerminating
    - ProjectPhaseFailed
...
host: localhost:8090
info:
  contact: {}
  description: Self-hosted Railway-style deploy platform for Kubernetes. Manages projects,
    apps, deployments, env vars, secrets, domains, and git providers.
  title: Mortise API
  version: 0.1.0
paths:
...
  /projects:
    get:
      description: Returns projects the caller has access to. Admins see all; members
        see only their projects.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.projectResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - BearerAuth: []
      summary: List projects
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Creates a new Project. Admin-only.
      parameters:
      - description: Project details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.createProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.projectResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - BearerAuth: []
      summary: Create a project
      tags:
      - projects
  /projects/{project}:
    delete:
      description: Deletes a Project. The controller tears down the backing namespace.
        Admin-only.
      parameters:
      - description: Project name
        in: path
        name: project
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - BearerAuth: []
      summary: Delete a project
      tags:
      - projects
    get:
      description: Returns a single Project by name
      parameters:
      - description: Project name
        in: path
        name: project
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.projectResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - BearerAuth: []
      summary: Get a project
      tags:
      - projects
...
  /projects/{project}/apps/{app}/connect:
    post:
      description: Start a reverse proxy listener on an auto-allocated local port
        for the app. Returns the existing proxy URL if already running.
      parameters:
      - description: Project name
        in: path
        name: project
        required: true
        type: string
      - description: App name
        in: path
        name: app
        required: true
        type: string
      - description: Environment name (defaults to first env)
        in: query
        name: environment
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.appProxyEntry'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - BearerAuth: []
      summary: Connect to an app via local proxy
      tags:
      - proxy
...
  /projects/{project}/apps/{app}/disconnect:
    post:
      description: Stop the reverse proxy listener for an app and free the allocated
        port.
      parameters:
      - description: Project name
        in: path
        name: project
        required: true
        type: string
      - description: App name
        in: path
        name: app
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - BearerAuth: []
      summary: Disconnect an app's local proxy
      tags:
      - proxy
...
securityDefinitions:
  BearerAuth:
    description: JWT bearer token from POST /api/auth/login
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"